requirements.txt
streamlit==1.39.0
supabase==2.7.4
pandas==2.2.2
plotly==5.24.1
python-dotenv==1.0.1
altair==5.4.1
numpy==1.26.4
python-dateutil==2.9.0
pytz==2024.2
streamlit-extras==0.4.7
tabulate==0.9.0
